+++ SpatialGDK/Source/SpatialGDKEditorToolbar/Public/SpatialGDKEditorToolbarCommands.h
@@ -24,13 +24,22 @@ public:
 	TSharedPtr<FUICommandInfo> CreateSpatialGDKSchemaFull;
 	TSharedPtr<FUICommandInfo> DeleteSchemaDatabase;
 	TSharedPtr<FUICommandInfo> CreateSpatialGDKSnapshot;
-	TSharedPtr<FUICommandInfo> StartSpatialDeployment;
+	TSharedPtr<FUICommandInfo> StartNoAutomaticConnection;
+	TSharedPtr<FUICommandInfo> StartLocalSpatialDeployment;
+	TSharedPtr<FUICommandInfo> StartCloudSpatialDeployment;
 	TSharedPtr<FUICommandInfo> StopSpatialDeployment;
 	TSharedPtr<FUICommandInfo> LaunchInspectorWebPageAction;
 	
 	TSharedPtr<FUICommandInfo> OpenCloudDeploymentWindowAction;
 	TSharedPtr<FUICommandInfo> OpenLaunchConfigurationEditorAction;
+	TSharedPtr<FUICommandInfo> EnableBuildClientWorker;
+	TSharedPtr<FUICommandInfo> EnableBuildSimulatedPlayer;
 
 	TSharedPtr<FUICommandInfo> StartSpatialService;
 	TSharedPtr<FUICommandInfo> StopSpatialService;
+	TSharedPtr<FUICommandInfo> EnableSpatialNetworking;
+	TSharedPtr<FUICommandInfo> GDKEditorSettings;
+	TSharedPtr<FUICommandInfo> NoAutomaticConnection;
+	TSharedPtr<FUICommandInfo> LocalDeployment;
+	TSharedPtr<FUICommandInfo> CloudDeployment;
 };